requirements.txt
flask>=2.0
yarl
requests
Flask-SQLAlchemy
psycopg2-binary>=2.9.0
Flask-Admin
flask-assets
pyscss
flask-wtf
PyYaml